Minor cleanup; move ordering to its own file
